refatora componentes da pagina relatorios
