+++ src/components/screens/Cadastros/CadastroHorarioComboAlimentacao/style.scss
@@ -141,7 +141,7 @@
   border: 1px solid #dadada !important;
 }
 
-.check-item {
+.checado-item-ok {
   height: 50%;
   padding: 0px 3px;
   border: 1px solid #52c41a;
