+++ src/components/screens/Relatorios/components/BuscaPorPeriodo/style.scss
@@ -1,0 +1,42 @@
+@import "../../../../../styles/variables.scss";
+
+.grid-container {
+  display: grid;
+  grid-template-columns: 59% 39%;
+  column-gap: 2%;
+  grid-template-rows: auto;
+  .container-left {
+    .container-left-bottom {
+      margin: 1em 0 0 0;
+      display: grid;
+      grid-template-columns: 30% 30% 36%;
+      column-gap: 2%;
+    }
+  }
+  label {
+    font-style: normal;
+    font-weight: bold;
+    font-size: 13px;
+    line-height: 165%;
+    color: #42474a;
+    margin: 0 0 0.5em 0 !important;
+  }
+  .container-right {
+    .container-right-bottom {
+      margin: 1em 0 0 0;
+      display: grid;
+      grid-template-columns: 72% 24%;
+      column-gap: 4%;
+      .filtrar {
+        display: flex;
+        justify-content: flex-end;
+      }
+    }
+  }
+}
+
+div.datepicker {
+  div.react-datepicker-wrapper {
+    width: 150px !important;
+  }
+}
