+++ src/components/screens/Relatorios/components/CardsComBandeira/style.scss
@@ -1,0 +1,100 @@
+@import "../../../../../styles/variables.scss";
+.cards-with-flag {
+  .fonte-grande {
+    font-style: normal;
+    font-weight: bold;
+    font-size: 26px;
+    line-height: 30px;
+    text-align: center;
+    color: #42474a;
+  }
+  .alinha-centro {
+    text-align: center;
+  }
+  .fonte-titulo {
+    font-family: Roboto;
+    font-style: normal;
+    font-weight: bold;
+    font-size: 18px;
+    line-height: 165%;
+    color: #42474a;
+  }
+  .card-solicitacoes {
+    display: grid;
+    grid-template-rows: auto;
+    border: 1px solid #eaedf3;
+    border-radius: 4px;
+    width: 262.98px;
+    height: 145px;
+    padding: 0.7em;
+    .header-card {
+      display: grid;
+      grid-template-columns: 20% 50% 30%;
+      max-height: 50px;
+      height: 44px;
+      .card-cancelled {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        background: rgba(228, 241, 255, 0.37);
+        box-shadow: 0px 5px 20px rgba(0, 0, 0, 0.03);
+        border-radius: 4px;
+        color: $black;
+      }
+      .card-denied {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        background: rgba(228, 241, 255, 0.37);
+        box-shadow: 0px 5px 20px rgba(0, 0, 0, 0.03);
+        border-radius: 4px;
+        color: $red;
+      }
+      .card-authorized {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        background: rgba(228, 241, 255, 0.37);
+        box-shadow: 0px 5px 20px rgba(0, 0, 0, 0.03);
+        border-radius: 4px;
+        color: $green;
+      }
+      .card-pending {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        background: rgba(228, 241, 255, 0.37);
+        box-shadow: 0px 5px 20px rgba(0, 0, 0, 0.03);
+        border-radius: 4px;
+        color: $orange;
+      }
+      .titulo-texto {
+        font-weight: 500;
+        font-size: 16px;
+        line-height: 19px;
+        display: flex;
+        align-items: center;
+        justify-content: flex-start;
+        color: #42474a;
+      }
+      .bandeira {
+        background: #c9251b;
+        color: #ffffff;
+        font-weight: bold;
+        font-size: 12px;
+        display: flex;
+        align-items: center;
+        &:before {
+          content: " ";
+          display: flex;
+          justify-content: space-around;
+          vertical-align: middle;
+          margin-right: 10px;
+          border-top: 22px solid transparent;
+          border-bottom: 22px solid transparent;
+          border-left: 20px solid #fff;
+        }
+      }
+    }
+  }
+}
