+++ src/components/screens/Relatorios/components/GraficoEvolucao/style.scss
@@ -1,0 +1,40 @@
+@import "../../../../../styles/variables.scss";
+
+div.evolution {
+  h2 {
+    font-weight: bold;
+    font-size: 18px;
+    color: $blackInput;
+  }
+  div.graph-and-legend {
+    padding: 20px;
+    border: 1px solid $grayDisabled;
+    div.graph {
+      border-right: 1px solid $grayDisabled;
+    }
+  }
+  .legend-unit {
+    .title {
+      display: inline-flex;
+      font-size: 14px;
+      color: $blackInput;
+      div.circle {
+        width: 10px;
+        height: 10px;
+        border-radius: 50%;
+        margin-top: 5px;
+        margin-right: 12px;
+      }
+      div.text {
+        margin-top: 3px;
+      }
+      &.small {
+        font-size: 11px;
+      }
+    }
+    .total {
+      font-size: 20px;
+      color: $blackSubtle;
+    }
+  }
+}
