+++ src/components/screens/Relatorios/components/ResultadoFiltro/style.scss
@@ -1,120 +1,6 @@
-@import "../../../styles/variables.scss";
+@import "../../../../../styles/variables.scss";
 
-div.evolution {
-  h2 {
-    font-weight: bold;
-    font-size: 18px;
-    color: $blackInput;
-  }
-  div.graph-and-legend {
-    padding: 20px;
-    border: 1px solid $grayDisabled;
-    div.graph {
-      border-right: 1px solid $grayDisabled;
-    }
-  }
-  .legend-unit {
-    .title {
-      display: inline-flex;
-      font-size: 14px;
-      color: $blackInput;
-      div.circle {
-        width: 10px;
-        height: 10px;
-        border-radius: 50%;
-        margin-top: 5px;
-        margin-right: 12px;
-      }
-      div.text {
-        margin-top: 3px;
-      }
-      &.small {
-        font-size: 11px;
-      }
-    }
-    .total {
-      font-size: 20px;
-      color: $blackSubtle;
-    }
-  }
-}
-
-.grid-container {
-  display: grid;
-  grid-template-columns: 59% 39%;
-  column-gap: 2%;
-  grid-template-rows: auto;
-
-  .container-left {
-    .container-left-bottom {
-      margin: 1em 0 0 0;
-      display: grid;
-      grid-template-columns: 30% 30% 36%;
-      column-gap: 2%;
-    }
-  }
-
-  label {
-    font-style: normal;
-    font-weight: bold;
-    font-size: 13px;
-    line-height: 165%;
-    color: #42474a;
-    margin: 0 0 0.5em 0 !important;
-  }
-
-  .container-right {
-    .container-right-bottom {
-      margin: 1em 0 0 0;
-      display: grid;
-      grid-template-columns: 72% 24%;
-      column-gap: 4%;
-
-      .filtrar {
-        display: flex;
-        justify-content: flex-end;
-      }
-    }
-  }
-}
-
-div.datepicker {
-  span.required-asterisk {
-    color: $red;
-    padding-right: 5px;
-  }
-  .react-datepicker__day--selected,
-  .react-datepicker__day--keyboard-selected {
-    background-color: $green;
-  }
-  .react-datepicker__close-icon {
-    display: none;
-  }
-  .fa-calendar-alt {
-    z-index: 0;
-    cursor: text;
-    padding-top: 12px;
-    color: $green;
-    position: relative;
-  }
-  div.react-datepicker-wrapper {
-    width: 150px !important;
-    margin-right: -30px;
-    .react-datepicker__input-container {
-      input {
-        height: 38px;
-        font-size: 14px;
-        color: $blackInput;
-        &.invalid-field:not(:focus) {
-          box-shadow: none;
-          border-color: $red;
-        }
-      }
-    }
-  }
-}
-
-.realtorio-filtro {
+.relatorio-filtro {
   header {
     font-style: normal;
     font-weight: normal;
@@ -309,16 +195,13 @@ div.datepicker {
   justify-content: flex-end;
   align-items: center;
   font-size: 12px;
-
   .pagina {
     padding: 0 0.5em;
     border: 1px solid #d1d3e2;
     border-radius: 4px;
     margin: 0 0.2em;
-
     cursor: pointer;
   }
-
   .atual {
     background-color: #d1d3e2;
     opacity: 0.5;
