+++ src/configs/helper.js
@@ -39,7 +39,7 @@ export const permissoes = () => {
 export const relatorios = () => {
   switch (localStorage.getItem("tipo_perfil")) {
     case TIPO_PERFIL.DIRETORIA_REGIONAL:
-      return RelatoriosPage.RelatoriosEscola;
+      return RelatoriosPage.RelatoriosDRE;
     case TIPO_PERFIL.GESTAO_ALIMENTACAO_TERCEIRIZADA:
       return RelatoriosPage.RelatoriosEscola;
     case TIPO_PERFIL.TERCEIRIZADA:
