+++ src/pages/Relatorios/RelatoriosPage.jsx
@@ -3,7 +3,11 @@ import Breadcrumb from "../../components/Shareable/Breadcrumb";
 import Relatorios from "../../components/screens/Relatorios";
 import Page from "../../components/Shareable/Page/Page";
 import { HOME } from "../../constants/config.constants";
-import { ESCOLA } from "../../configs/constants";
+import { ESCOLA, DRE } from "../../configs/constants";
+import {
+  getEvolucaoSolicitacoesEscola,
+  getResumoTotaisPorMesEscola
+} from "../../services/relatorios.service";
 
 class RelatoriosBase extends React.Component {
   render() {
@@ -22,4 +26,13 @@ class RelatoriosBase extends React.Component {
 }
 
 // Escola
-export const RelatoriosEscola = () => <RelatoriosBase visao={ESCOLA} />;
+export const RelatoriosEscola = () => (
+  <RelatoriosBase
+    visao={ESCOLA}
+    getEvolucaoSolicitacoes={getEvolucaoSolicitacoesEscola}
+    getResumoTotaisPorMes={getResumoTotaisPorMesEscola}
+  />
+);
+
+// DRE
+export const RelatoriosDRE = () => <RelatoriosBase visao={DRE} />;
